Product editor media query responsiveness
