+++ packages/fdg_web_admin/lib/src/product_unit_localization.dart
@@ -1,7 +1,7 @@
 import 'package:fdg_web_admin/src/ui/products/product_unit.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:fdg_web_admin/src/ui/products/editor/product_editor_dialog.dart';
+import 'package:fdg_web_admin/src/ui/products/editor/product_editor.dart';
 import 'package:fdg_web_admin/src/fdg_products_locale_keys.dart';
